+++ webpack.config.js
@@ -1,0 +1,17 @@
+const { VueLoaderPlugin } = require("vue-loader");
+module.exports = {
+  module: {
+    rules: [
+      {
+        test: /\.vue$/,
+        loader: "vue-loader",
+      },
+    ],
+  },
+  plugins: [new VueLoaderPlugin()],
+  output: {
+    filename: "tvdemo.js",
+    library: "Tvdemo",
+    libraryTarget: "umd",
+  },
+};